Calendar Implementation of automatic disabling  the page links of empty pages:
This implementation works for pagy_bootstrap_nav.
For AR it rely on the pagy_calendar_counts(collection, filter_series) implemented in the application. That method should calculate the count for each passed filter and return it as a hash.
Search frameworks have easy aggregations and could return the counts with the result itself. That requires an alternative way to set the @counts.

TODO:
- Implement it for all the *nav
- Implement it for all the nav_js* and in javascripts/pagy.js
- Find an efficient way to get the counts in AR
- Find a way to use the counts to add a feedback for each page (tooltip?)
- Tests
- Docs

Notice: counts allow to have tooltips for the count on each page.